// web/src/app/model/join-condition-operator.ts
export enum JoinConditionOperator {
  EQUALS = 'EQUALS',
  GREATER_THAN = 'GT',
  LESS_THAN = 'LT',
  GREATER_THAN_EQUALS = 'GT_EQUALS',
  LESS_THAN_EQUALS = 'LT_EQUALS',
  NOT_EQUALS = 'NOT_EQUALS',

  // string operators
  STARTS_WITH = 'STARTS_W',
  ENDS_WITH = 'ENDS_W',
  CONTAINS = 'CONTAINS',
}